rename: 엔티티 클래스들과 JPA 설정 클래스 domain 모듈로 이동
- RoleType 이넘 패키지 변경으로 인한 일부 수정 포함
- 엔티티 관련 테스트도 이동 및 테스트 코드에서 불필요한 참조 제거
