+++ dailyquest_domain/src/main/java/dailyquest/common/BaseLogEntity.java
