+++ dailyquest_domain/src/main/java/dailyquest/common/BaseTimeEntity.kt
