+++ dailyquest_domain/src/main/java/dailyquest/quest/entity/DetailQuest.kt
