+++ dailyquest_domain/src/main/java/dailyquest/user/entity/ProviderType.java
